Update 1.4
Adding UI features such as score and time as well as a game over feature.
Integrated dependency injection for certain scripts.
Added a particle system when an apple is picked up for visual flair.
Cinemachine implemented and when an apple is picked up, changing the FOV so it appears zoomed out and zooming in when hitting obstacles.
Created certain ranges for min and max speed.