Update 1.7 - Final Build
Implementation of:
- Music for overall game
- Tweaked text designs
- Added the score to the Game Over text
- Added a feature where the player passing the checkpoint, decreases the obstacle spawner forcing more obstacles to have the player eventually slow down and lose the game. Essentially game balancing/ adding difficulty
- Tweaked settings to physics/gravity in project settings and fog. Changed spawn percentages to apple in general and spawn percentages to checkpoint prefab.

Once finished, made a build of the game for Windows

NOTE: There is a slight bug with gaps between some chunks, seems to be happening 'randomly'. Seems to not suggest the math when generating them or the size or positions of the prefabs. Possibly a rendering issue as the player doesn't fall, meaning there is a collider or some sort of object. Seems to have just occurred towards end of project